* Abstract our custom action bar implementation to be a standalone widget. This affords us the ability to use it for both the action bar and an action mode at the same time. It also would allow someone to use the widget without fully utilizing the automatic action bar abstraction. * Rework themes to mimic those of the Android 3.0+ native themes except as Theme.Sherlock and Theme.Sherlock.Light. These automatically inherit from the appropriate parent themes based on platform.
Still need to mimic Android's MenuItemImpl which holds weak references to its views (in our case, ActionBarWatson$Item) and automatically passes changes to certain properties to be reflected instantly. This will allow for holding on to instances of MenuBuilder (as Menu) to update the action bar items asynchronously of interaction.