Create a proper plugin system.
Proper plugin-based system for selecting the best implementation of the ActionBarSherlock abstract class. Internally we provide solutions for API 7+ and API 14+ but now we enable future work to be done on providing an API 3+ implementation as well as a TVDPI implementation.

This also comes with the added bonus of being able to expose the implementation of ActionBarSherlock through the `getSherlock()` method (the interface through which static attachment occurs). With this certain window-based properties can be set without requiring explicit methods in the activity base classes. Closes #278.