Move support-lib plugin into library.
The vast majority of users that are interested in this library will also be using the official support library. If they're not, they most likely should be anyways.

This comes with the added bonus of re-enabling the transaction parameter on the tab callbacks. We have to do a bit of magic when converting the native callbacks but it operates under the same callback logic as we do so the assumptions should be plenty safe.